refactor: make _reconcile package public
