+++ src/fennflow/exceptions/__init__.py
@@ -12,7 +12,6 @@ __all__ = [
     "StoragePrefixIsNoneException",
 ]
 from fennflow._base_exception import FennFlowException
-from fennflow._reconciler.exceptions import ReconcileFailedException
 from fennflow.backends.exceptions import (
     BaseBackendException,
     RecordAlreadyExistsException,
@@ -26,3 +25,4 @@ from fennflow.files.exceptions import (
     MediaTypeCannotBeGuessedException,
     StoragePrefixIsNoneException,
 )
+from fennflow.reconciler.exceptions import ReconcileFailedException
