+++ tests/conftest.py
@@ -1,10 +1,10 @@
 import pytest
 
 from fennflow import ConfigDict
-from fennflow._reconciler._orchestrator import ReconcileOrchestrator
 from fennflow.backends import InMemoryBackend, InMemoryBackendConfig
 from fennflow.connectors import InMemoryConnector, InMemoryConnectorConfig
 from fennflow.files import TextContent
+from fennflow.reconciler._orchestrator import ReconcileOrchestrator
 from fennflow.repositories import (
     DeleteRepository,
     GetRepository,
